--- KitBot-2021/.wpilib/wpilib_preferences.json
@@ -1,6 +1,0 @@
-{
-    "enableCppIntellisense": false,
-    "currentLanguage": "java",
-    "projectYear": "2021",
-    "teamNumber": 967
-}
