--- KitBot-2021/src/main/deploy/robotMap.properties
@@ -1,8 +1,0 @@
-rightMaster = 1
-rightSlave = 2
-leftMaster = 3
-leftSlave = 4
-linearActuator = 6
-turretRot = 5
-leftFlywheel = 7
-rightFlywheel = 8
