--- KitBot-2021/src/main/deploy/values.properties
@@ -1,9 +1,0 @@
-driveDeadband = 0.1
-driveMinPower = 0.05
-driveLowCutoff = 1.0
-driveMidPower = 1.0
-
-turnDeadband = 0.07
-turnMinPower = 0.03
-turnLowCutoff = 0.5
-turnMidPower = 0.4
