--- KitBot-2021/src/main/java/frc/robot/commands/ArcadeDriveCommand.java
@@ -1,50 +1,0 @@
-/*----------------------------------------------------------------------------*/
-/* Copyright (c) 2019 FIRST. All Rights Reserved.                             */
-/* Open Source Software - may be modified and shared by FRC teams. The code   */
-/* must be accompanied by the FIRST BSD license file in the root directory of */
-/* the project.                                                               */
-/*----------------------------------------------------------------------------*/
-
-package frc.robot.commands;
-
-import edu.wpi.first.wpilibj2.command.CommandBase;
-import frc.robot.IO;
-import frc.robot.subsystems.SubsystemsInstance;
-
-public class ArcadeDriveCommand extends CommandBase {
-  private SubsystemsInstance inst;
-  private IO io;
-
-  /**
-   * Creates a new ArcadeDriveCommand.
-   */
-  public ArcadeDriveCommand() {
-    // Use addRequirements() here to declare subsystem dependencies.
-    inst = SubsystemsInstance.getInstance();
-    addRequirements(inst.m_driveSubsystem);
-
-    io = IO.getInstance();
-  }
-
-  // Called when the command is initially scheduled.
-  @Override
-  public void initialize() {
-  }
-
-  // Called every time the scheduler runs while the command is scheduled.
-  @Override
-  public void execute() {
-    inst.m_driveSubsystem.arcadeDrive(-io.getDriverController().getRightStickX(), -io.getDriverController().getLeftStickY());
-  }
-
-  // Called once the command ends or is interrupted.
-  @Override
-  public void end(boolean interrupted) {
-  }
-
-  // Returns true when the command should end.
-  @Override
-  public boolean isFinished() {
-    return false;
-  }
-}
