--- KitBot-2021/src/main/java/frc/robot/utils/Utils.java
@@ -1,9 +1,0 @@
-package frc.robot.utils;
-
-public class Utils {
-
-    //method that returns the input if it is outside the deadband parameter, or 0 if it is inside
-    public static double deadband(double x, double d) {
-        return ((x > d) || (x < -d)) ? x : 0;
-    }
-}
