--- KitBot-2021/src/main/java/frc/robot/values/LookupType.java
@@ -1,6 +1,0 @@
-package frc.robot.values;
-
-public enum LookupType {
-    DOUBLE_LINEAR,
-    QUADRATIC
-}
