--- KitBot/.vs/VSWorkspaceState.json
@@ -1,6 +1,0 @@
-{
-  "ExpandedNodes": [
-    ""
-  ],
-  "PreviewInSolutionExplorer": false
-}
