+++ KitBot/.wpilib/wpilib_preferences.json
@@ -1,6 +1,6 @@
 {
     "enableCppIntellisense": false,
     "currentLanguage": "java",
-    "projectYear": "2020",
+    "projectYear": "2021",
     "teamNumber": 967
 }
