+++ KitBot/src/main/java/frc/robot/utils/controls/XBoxController.java
@@ -7,6 +7,7 @@ import edu.wpi.first.wpilibj2.command.button.POVButton;
 import edu.wpi.first.wpilibj2.command.button.JoystickButton;
 import edu.wpi.first.wpilibj2.command.Command;
 
+
 public class XBoxController extends Joystick {
     public static final int LEFT_X_AXIS = 0;
     public static final int LEFT_Y_AXIS = 1;
