+++ KitBot/vendordeps/navx_frc.json
@@ -1,17 +1,17 @@
 {
     "fileName": "navx_frc.json",
     "name": "KauaiLabs_navX_FRC",
-    "version": "3.1.400",
+    "version": "4.0.414",
     "uuid": "cb311d09-36e9-4143-a032-55bb2b94443b",
     "mavenUrls": [
         "https://repo1.maven.org/maven2/"
     ],
-    "jsonUrl": "https://www.kauailabs.com/dist/frc/2020/navx_frc.json",
+    "jsonUrl": "https://www.kauailabs.com/dist/frc/2021/navx_frc.json",
     "javaDependencies": [
         {
             "groupId": "com.kauailabs.navx.frc",
             "artifactId": "navx-java",
-            "version": "3.1.400"
+            "version": "4.0.414"
         }
     ],
     "jniDependencies": [],
@@ -19,7 +19,7 @@
         {
             "groupId": "com.kauailabs.navx.frc",
             "artifactId": "navx-cpp",
-            "version": "3.1.400",
+            "version": "4.0.414",
             "headerClassifier": "headers",
             "sourcesClassifier": "sources",
             "sharedLibrary": false,
@@ -27,7 +27,8 @@
             "skipInvalidPlatforms": true,
             "binaryPlatforms": [
                 "linuxathena",
-                "linuxraspbian"
+                "linuxraspbian",
+                "windowsx86-64"
             ]
         }
     ]
